fix: according to review feedback
