+++ samples/datatypes.js
@@ -607,151 +607,9 @@ async function queryWithTimestamp(instanceId, databaseId, projectId) {
   // [END spanner_query_with_timestamp_parameter]
 }
 
-async function addNumericColumn(instanceId, databaseId, projectId) {
-  // [START spanner_add_numeric_column]
-  // Imports the Google Cloud client library.
-  const {Spanner} = require('@google-cloud/spanner');
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const projectId = 'my-project-id';
-  // const instanceId = 'my-instance';
-  // const databaseId = 'my-database';
-
-  // Creates a client
-  const spanner = new Spanner({
-    projectId: projectId,
-  });
-
-  // Gets a reference to a Cloud Spanner instance.
-  const instance = spanner.instance(instanceId);
-  const database = instance.database(databaseId);
-
-  const request = ['ALTER TABLE Venues ADD COLUMN Revenue NUMERIC'];
-
-  // Alter existing table to add a column.
-  const [operation] = await database.updateSchema(request);
-
-  console.log(`Waiting for operation on ${databaseId} to complete...`);
-
-  await operation.promise();
-
-  console.log(
-    `Added Revenue column to Venues table in database ${databaseId}.`
-  );
-  // [END spanner_add_numeric_column]
-}
-
-async function updateWithNumericData(instanceId, databaseId, projectId) {
-  // [START spanner_update_data_with_numeric]
-  // Imports the Google Cloud client library.
-  const {Spanner} = require('@google-cloud/spanner');
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const projectId = 'my-project-id';
-  // const instanceId = 'my-instance';
-  // const databaseId = 'my-database';
-
-  // Creates a client.
-  const spanner = new Spanner({
-    projectId: projectId,
-  });
-
-  // Gets a reference to a Cloud Spanner instance and database.
-  const instance = spanner.instance(instanceId);
-  const database = instance.database(databaseId);
-
-  // Instantiate Spanner table objects.
-  const venuesTable = database.table('Venues');
-
-  const data = [
-    {
-      VenueId: '4',
-      Revenue: Spanner.numeric('35000'),
-      LastUpdateTime: 'spanner.commit_timestamp()',
-    },
-    {
-      VenueId: '19',
-      Revenue: Spanner.numeric('104500'),
-      LastUpdateTime: 'spanner.commit_timestamp()',
-    },
-    {
-      VenueId: '42',
-      Revenue: Spanner.numeric('99999999999999999999999999999.99'),
-      LastUpdateTime: 'spanner.commit_timestamp()',
-    },
-  ];
-
-  // Updates rows in the Venues table.
-  try {
-    await venuesTable.update(data);
-    console.log('Updated data.');
-  } catch (err) {
-    console.error('ERROR:', err);
-  } finally {
-    // Close the database when finished.
-    database.close();
-  }
-  // [END spanner_update_data_with_numeric]
-}
-
-async function queryWithNumeric(instanceId, databaseId, projectId) {
-  // [START spanner_query_with_numeric_parameter]
-  // Imports the Google Cloud client library.
-  const {Spanner} = require('@google-cloud/spanner');
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const projectId = 'my-project-id';
-  // const instanceId = 'my-instance';
-  // const databaseId = 'my-database';
-
-  // Creates a client
-  const spanner = new Spanner({
-    projectId: projectId,
-  });
-
-  // Gets a reference to a Cloud Spanner instance and database.
-  const instance = spanner.instance(instanceId);
-  const database = instance.database(databaseId);
-
-  const fieldType = {
-    type: 'numeric',
-  };
-
-  const exampleNumeric = Spanner.numeric('100000');
-
-  const query = {
-    sql: `SELECT VenueId, VenueName, Revenue FROM Venues
-            WHERE Revenue < @revenue`,
-    params: {
-      revenue: exampleNumeric,
-    },
-    types: {
-      revenue: fieldType,
-    },
-  };
-
-  // Queries rows from the Venues table.
-  try {
-    const [rows] = await database.run(query);
-
-    rows.forEach(row => {
-      const json = row.toJSON();
-      console.log(`VenueId: ${json.VenueId}, Revenue: ${json.Revenue.value}`);
-    });
-  } catch (err) {
-    console.error('ERROR:', err);
-  } finally {
-    // Close the database when finished.
-    database.close();
-  }
-  // [END spanner_query_with_numeric_parameter]
-}
+const {addNumericColumn} = require('./numeric-add-column');
+const {updateWithNumericData} = require('./numeric-update-data');
+const {queryWithNumericParameter} = require('./numeric-query-parameter');
 
 require('yargs')
   .demand(1)
@@ -837,11 +695,15 @@ require('yargs')
       )
   )
   .command(
-    'queryWithNumeric <instanceName> <databaseName> <projectId>',
+    'queryWithNumericParameter <instanceName> <databaseName> <projectId>',
     "Query data from the sample 'Venues' table with a NUMERIC datatype.",
     {},
     opts =>
-      queryWithNumeric(opts.instanceName, opts.databaseName, opts.projectId)
+      queryWithNumericParameter(
+        opts.instanceName,
+        opts.databaseName,
+        opts.projectId
+      )
   )
   .example(
     'node $0 createVenuesTable "my-instance" "my-database" "my-project-id"'
@@ -866,7 +728,7 @@ require('yargs')
     'node $0 updateWithNumericData "my-instance" "my-database" "my-project-id"'
   )
   .example(
-    'node $0 queryWithNumeric "my-instance" "my-database" "my-project-id"'
+    'node $0 queryWithNumericParameter "my-instance" "my-database" "my-project-id"'
   )
   .wrap(120)
   .recommendCommands()
